[Artist] Embeddings using GMM (hyperparam search for each artist)
We do hyperparameter search over number of components and covariance type
for a GMM for each artist. The fit is the log likelihood on a
test subset. For each artist, we fit it N times and take the
hyperparameter setting with the best average fit across the N runs.

Test plan:
- python artist_embeddings.py, which ran:
  - TSNE visualization
  - calculate_and_save_artist_gmm_and_embs()
  - print_artist_to_num_nodes()

Some other notes:
- Since the log likelihood is the only way to evaluate fit on the
high dimensional embeddings, I also ran GMM on TSNE embeddings
to see if at least the number of modes was approximately the same.
This was the case.
- I didn't include the 'tied' or 'full' covariance type in the
hyperparameter search. Trying it on a bunch of artists indicated
it always overfit, which isn't too surprising as there aren't
that many data points (paintings) per artist.
- I also ran the TSNE embeddings and visualized them for a few
artists to sanity check the painting embeddings. Seems to make sense.
  - I also ran it separately on content only and style only. It seems
 like the style only does better actually, at least visually.

Stats:
Num artists with 1 mode: 112
Num artists with 2 modes: 55
Num artists with 3 modes: 36
Num artists with 4 modes: 17
Num artists with 5 modes: 21
Num artists with 6 modes: 12
Num artists with 7 modes: 16
Num artists with 8 modes: 11
Num artists with 9 modes: 7
Num artists with 10 modes: 11